Rather than waiting for active uploads - monitor the current storage size and keep trying while there's still time left to upload whatever is there. This is equivalent to v1 behavior.

+++ lib/include/aria/IOfflineStorage.hpp
@@ -247,7 +247,7 @@ namespace ARIASDK_NS_BEGIN {
         /// Get current Db size returned. Called from the internal worker thread.
         /// </remarks>
         /// <returns>Value of the requested DB size</returns>
-        virtual unsigned GetSize() = 0;
+        virtual size_t GetSize() = 0;
 
         /// <summary>
         /// Get Vector of records from DB
