+++ lib/offline/OfflineStorage_SQLite.hpp
@@ -39,7 +39,7 @@ namespace ARIASDK_NS_BEGIN {
         virtual void ReleaseRecords(std::vector<StorageRecordId> const& ids, bool incrementRetryCount, HttpHeaders headers, bool& fromMemory) override;
         virtual bool StoreSetting(std::string const& name, std::string const& value) override;
         virtual std::string GetSetting(std::string const& name) override;
-        virtual unsigned GetSize() override;
+        virtual size_t GetSize() override;
         virtual std::vector<StorageRecord>* GetRecords(bool shutdown, EventLatency minLatency = EventLatency_Normal, unsigned maxCount = 0) override;
         virtual bool ResizeDb() override;
